+++ pol-core/bscript/compiler/file/SourceLocation.h
@@ -20,8 +20,8 @@ class SourceFileIdentifier;
 
 struct Position
 {
-  unsigned short line_number;
-  unsigned short character_column;  // 1-based on line, as seen in an editor
+  size_t line_number;
+  size_t character_column;  // 1-based on line, as seen in an editor
   size_t token_index;
 };
 
@@ -31,7 +31,7 @@ struct Range
   Position end;
   bool contains( const Position& ) const;
   bool contains( const Range& ) const;
-  bool contains( unsigned short line_number, unsigned short character_column ) const;
+  bool contains( size_t line_number, size_t character_column ) const;
 
   Range( const antlr4::ParserRuleContext& );
   Range( const antlr4::tree::TerminalNode& ctx );
@@ -42,8 +42,7 @@ struct Range
 class SourceLocation
 {
 public:
-  SourceLocation( const SourceFileIdentifier*, unsigned short line_number,
-                  unsigned short character_column );
+  SourceLocation( const SourceFileIdentifier*, size_t line_number, size_t character_column );
   SourceLocation( const SourceFileIdentifier*, const Range& );
   SourceLocation( const SourceFileIdentifier*, const antlr4::ParserRuleContext& );
   SourceLocation( const SourceFileIdentifier*, const antlr4::tree::TerminalNode& );
